+++ DV1572/DV1572/ourEngine/interface/Window.cpp
@@ -899,7 +899,7 @@ LRESULT Window::WndProc(UINT uMsg, WPARAM wParam, LPARAM lParam)
 			PostQuitMessage(0);
 		
 		Input::m_keys[wParam] = true;
-		Input::lastPressed = wParam;
+		Input::lastPressed = static_cast<int>(wParam);
 		break;
 	case WM_KEYUP:
 		Input::m_keys[wParam] = false;
